+++ lib/payment_request/payment_request.rb
@@ -142,8 +142,9 @@ module StarkBank
       return [payment, 'boleto-payment'] if payment.is_a?(StarkBank::BoletoPayment)
       return [payment, 'utility-payment'] if payment.is_a?(StarkBank::UtilityPayment)
       return [payment, 'tax-payment'] if payment.is_a?(StarkBank::TaxPayment)
+      return [payment, 'darf-payment'] if payment.is_a?(StarkBank::DarfPayment)
 
-      raise(Exception('Payment must either be a Transfer, a Transaction, a BrcodePayment, BoletoPayment, a UtilityPayment, a TaxPayment or a hash.')) unless payment.is_a?(Hash)
+      raise(Exception('Payment must either be a Transfer, a Transaction, a BrcodePayment, BoletoPayment, a UtilityPayment, a TaxPayment, a DarfPayment or a hash.')) unless payment.is_a?(Hash)
 
       resource = {
         'transfer': StarkBank::Transfer.resource,
@@ -151,7 +152,8 @@ module StarkBank
         'brcode-payment': StarkBank::BrcodePayment.resource,
         'boleto-payment': StarkBank::BoletoPayment.resource,
         'utility-payment': StarkBank::UtilityPayment.resource,
-        'tax-payment': StarkBank::TaxPayment.resource
+        'tax-payment': StarkBank::TaxPayment.resource,
+        'darf-payment': StarkBank::DarfPayment.resource
       }[type.to_sym]
 
       payment = StarkBank::Utils::API.from_api_json(resource[:resource_maker], payment) unless resource.nil?