+++ test/starkbank/test_boleto_log.rb
@@ -29,9 +29,7 @@ describe(StarkBank::Boleto::Log, '#boleto/log#') do
         expect(ids).wont_include(log.id)
         ids << log.id
       end
-      if cursor.nil?
-        break
-      end
+      break if cursor.nil?
     end
     expect(ids.length).must_equal(10)
   end